+++ src/services/user.service.ts
@@ -1,0 +1,12 @@
+import { createClient } from "@/lib/utils/supabase/server";
+
+export async function getCurrentUser() {
+  const supabase = createClient();
+  const { data, error } = await supabase.auth.getUser();
+
+  if (error) {
+    return { error: error.message };
+  } else {
+    return { data };
+  }
+}
